Remove boilerplate
